refactor: admin/tags featureをserver/client/shared構成にリストラクチャ
