+++ admin/src/features/bills-edit/components/bill-tags-form.tsx
@@ -5,7 +5,7 @@ import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Label } from "@/components/ui/label";
-import type { Tag } from "@/features/tags/types";
+import type { Tag } from "@/features/tags/shared/types";
 import { updateBillTags } from "../actions/update-bill-tags";
 
 interface BillTagsFormProps {
